src/backend/requirements.txt
fastapi
uvicorn

azure-cosmos
azure-monitor-opentelemetry
azure-monitor-events-extension
azure-identity
python-dotenv
python-multipart
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-openai
opentelemetry-exporter-otlp-proto-http

semantic-kernel[azure]==1.29.0
azure-ai-projects
openai 
azure-ai-inference 
azure-search-documents 
azure-ai-evaluation

opentelemetry-exporter-otlp-proto-grpc

# Testing tools
pytest>=8.2,<9  # Compatible version for pytest-asyncio
pytest-asyncio==0.24.0
pytest-cov==5.0.0

